Make the passphrase generate new wallets on unlock
