+++ src/pages/HomePage.tsx
@@ -27,6 +27,7 @@ import Button from '../components/Button'
 import SideBar from '../components/HomePage/SideBar'
 import Input from '../components/Inputs/Input'
 import Select from '../components/Inputs/Select'
+import WalletPassphrase from '../components/Inputs/WalletPassphrase'
 import { FloatingPanel, Section } from '../components/PageComponents/PageContainers'
 import PanelTitle from '../components/PageComponents/PanelTitle'
 import Paragraph from '../components/Paragraph'
@@ -98,6 +99,7 @@ const Login = ({ walletNames, onLinkClick }: LoginProps) => {
   const [credentials, setCredentials] = useState({ walletName: '', password: '' })
   const { login } = useGlobalContext()
   const navigate = useNavigate()
+  const [passphrase, setPassphraseState] = useState('')
 
   const handleCredentialsChange = useCallback((type: 'walletName' | 'password', value: string) => {
     setCredentials((prev) => ({ ...prev, [type]: value }))
@@ -105,8 +107,15 @@ const Login = ({ walletNames, onLinkClick }: LoginProps) => {
 
   const handleLogin = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     e.preventDefault()
-    login(credentials.walletName, credentials.password, () => navigate('/wallet/overview'))
+    login(credentials.accountName, credentials.password, () => navigate('/wallet/overview'), passphrase)
   }
+
+  const onUpdatePassphrase = useCallback(
+    (e: ChangeEvent<HTMLInputElement>): void => {
+      setPassphraseState(e.target.value)
+    },
+    [setPassphraseState]
+  )
 
   return (
     <>
@@ -125,6 +134,12 @@ const Login = ({ walletNames, onLinkClick }: LoginProps) => {
           onChange={(e) => handleCredentialsChange('password', e.target.value)}
           value={credentials.password}
           id="password"
+        />
+        <WalletPassphrase
+          value={passphrase}
+          label="Optional passphrase"
+          onChange={onUpdatePassphrase}
+          isValid={passphrase.length > 0}
         />
       </SectionStyled>
       <SectionStyled inList>
